requirements.txt
black >= 24.10.0, < 25.0.0
click >= 8.1.8, < 9.0.0
loguru >= 0.7.3, < 1.0.0
microdot >= 2.0.7, < 3.0.0
pytest >= 8.3.4, < 9.0.0
python-can >= 4.5.0, < 5.0.0
pyserial >= 3.5.0, < 4.0.0
pytest >= 8.3.4, < 9.0.0
pyyaml >= 6.0.0, < 7.0.0
requests >= 2.32.3, < 3.0.0
ruff >= 0.9.7, < 0.10.0 # ruff is in early development, pin to minor version for now
